Separate scales, metrics, their builders, etc Move their to new metrics dir and namespace
