+++ arangod/Metrics/Fwd.h
@@ -1,0 +1,49 @@
+////////////////////////////////////////////////////////////////////////////////
+/// DISCLAIMER
+///
+/// Copyright 2014-2021 ArangoDB GmbH, Cologne, Germany
+/// Copyright 2004-2014 triAGENS GmbH, Cologne, Germany
+///
+/// Licensed under the Apache License, Version 2.0 (the "License");
+/// you may not use this file except in compliance with the License.
+/// You may obtain a copy of the License at
+///
+///     http://www.apache.org/licenses/LICENSE-2.0
+///
+/// Unless required by applicable law or agreed to in writing, software
+/// distributed under the License is distributed on an "AS IS" BASIS,
+/// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+/// See the License for the specific language governing permissions and
+/// limitations under the License.
+///
+/// Copyright holder is ArangoDB GmbH, Cologne, Germany
+///
+/// @author Valery Mironov
+////////////////////////////////////////////////////////////////////////////////
+#pragma once
+
+namespace arangodb::metrics {
+
+template <typename T>
+class Batch;
+
+class Counter;
+
+template <typename T>
+class Gauge;
+
+template <typename T>
+class FixScale;
+
+template <typename T>
+class LinScale;
+
+template <typename T>
+class LogScale;
+
+template <typename Scale>
+class Histogram;
+
+class MetricsFeature;
+
+}  // namespace arangodb::metrics
