+++ arangod/Metrics/MetricKey.h
@@ -1,0 +1,41 @@
+////////////////////////////////////////////////////////////////////////////////
+/// DISCLAIMER
+///
+/// Copyright 2014-2021 ArangoDB GmbH, Cologne, Germany
+/// Copyright 2004-2014 triAGENS GmbH, Cologne, Germany
+///
+/// Licensed under the Apache License, Version 2.0 (the "License");
+/// you may not use this file except in compliance with the License.
+/// You may obtain a copy of the License at
+///
+///     http://www.apache.org/licenses/LICENSE-2.0
+///
+/// Unless required by applicable law or agreed to in writing, software
+/// distributed under the License is distributed on an "AS IS" BASIS,
+/// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+/// See the License for the specific language governing permissions and
+/// limitations under the License.
+///
+/// Copyright holder is ArangoDB GmbH, Cologne, Germany
+///
+/// @author Kaveh Vahedipour
+////////////////////////////////////////////////////////////////////////////////
+#pragma once
+
+#include <cstddef>
+#include <string_view>
+
+namespace arangodb::metrics {
+
+struct MetricKey {
+  std::string_view name;
+  std::string_view labels;
+
+  MetricKey() noexcept;
+  MetricKey(std::string_view name) noexcept;
+  MetricKey(std::string_view name, std::string_view labels) noexcept;
+};
+
+bool operator<(MetricKey const& lhs, MetricKey const& rhs);
+
+}  // namespace arangodb::metrics
